requirement.txt
python_weather
kafka-python
datapackage